Trait orca::pipelines::Pipeline

pub trait Pipeline {
    // Required methods
    fn execute<'life0, 'life1, 'async_trait>(
        &'life0 self,
        target: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<PipelineResult>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn context(&mut self) -> &mut HashMap<String, String>;

    // Provided methods
    fn load_context<'life0, 'life1, 'async_trait, T>(
        &'life0 mut self,
        context: &'life1 T
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where T: Serialize + Sync + 'async_trait,
             Self: Sized + Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn load_record(&mut self, name: &str, record: Record) { ... }
}

Required Methods

fn execute<'life0, 'life1, 'async_trait>( &'life0 self, target: &'life1 str ) -> Pin<Box<dyn Future<Output = Result<PipelineResult>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Executes a given pipeline and produces an LLM response.

Returns

A Result containing a PipelineResult if successful or an error otherwise.

fn context(&mut self) -> &mut HashMap<String, String>

Retrieves the current context of the pipeline.

Returns

A mutable reference to a hashmap containing the current context.

Provided Methods

fn load_context<'life0, 'life1, 'async_trait, T>( &'life0 mut self, context: &'life1 T ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where T: Serialize + Sync + 'async_trait, Self: Sized + Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Sets the context for the current pipeline execution using a given data structure.

Parameters

context: A reference to a serializable data structure.

Examples

use orca::pipelines::Pipeline;
use orca::llm::openai::OpenAI;
use orca::pipelines::pipeline::LLMPipeline;
use std::collections::HashMap;

let client = OpenAI::new();
let mut pipeline = LLMPipeline::new(&client).with_template("my prompt", "Hello, {{name}}!");
let mut data = HashMap::new();
data.insert("name", "LLM");
pipeline.load_context(&data).await;

fn load_record(&mut self, name: &str, record: Record)

Loads a given record into the context of the LLM pipeline.

Parameters

name: The key/name for the record content in the context.

record: The actual record to load.