Trait orca::llm::LLM

pub trait LLM: Sync + Send {
    // Required method
    fn generate<'life0, 'async_trait>(
        &'life0 self,
        prompt: Box<dyn Prompt>
    ) -> Pin<Box<dyn Future<Output = Result<LLMResponse>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generate with context trait is used to execute an LLM using a context and a prompt template. The context is a previously created context using the Context struct. The prompt template is a previously created prompt template using the template! macro.

Required Methods

fn generate<'life0, 'async_trait>( &'life0 self, prompt: Box<dyn Prompt> ) -> Pin<Box<dyn Future<Output = Result<LLMResponse>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Generate a response from an LLM using a context and a prompt template.

Arguments

prompt - A prompt trait object.

Examples

This example uses the OpenAI chat models.

use orca::llm::LLM;
use orca::prompt::Prompt;
use orca::template;
use orca::llm::openai::OpenAI;
use orca::prompt::TemplateEngine;

#[tokio::main]
async fn main() {
   let prompt = template!(
      "my template",
      r#"
      {{#chat}}
      {{#user}}
      What is the capital of France?
      {{/user}}
      {{/chat}}
      "#
   );
   let client = OpenAI::new();
   let prompt = prompt.render("my template").unwrap();
   let response = client.generate(prompt).await.unwrap();
   assert!(response.to_string().to_lowercase().contains("paris"));
}