Struct orca::prompt::TemplateEngine

pub struct TemplateEngine {
    pub templates: HashMap<String, String>,
    /* private fields */
}

Represents a prompt engine that uses handlebars templates to render strings.

Fields
templates: HashMap<String, String>

Registered templates

Implementations

impl TemplateEngine

pub fn new() -> TemplateEngine

Creates a new TemplateEngine with the given prompt string.

Arguments

prompt - A string slice that holds the prompt template.

Example

use orca::prompt::TemplateEngine;
let prompt = TemplateEngine::new();

pub fn register_template(self, name: &str, template: &str) -> Self

pub fn get_template(&self, name: &str) -> Option<String>

pub fn add_to_template(&mut self, name: &str, new_template: &str)

Adds a new template to the prompt.

This function appends a new template to the existing prompt. The template should be a string slice that holds the template to be added. The template will be added to the prompt on a new line.

Arguments

template - A string slice that holds the template to be added.

Example

use orca::prompt::TemplateEngine;

let mut prompt = TemplateEngine::new().register_template("template", "Welcome!");
prompt.add_to_template("template", "Hello, world!");
assert_eq!(prompt.templates["template"], "Welcome!Hello, world!");

pub fn render(&self, name: &str) -> Result<Box<dyn Prompt>>

Renders a Handlebars template and returns the result as a Boxed trait object.

Arguments

data - A reference to the data to be used in the template rendering.

Returns

Returns a Boxed trait object that implements the Prompt trait.

Example

use serde_json::json;
use orca::prompt::TemplateEngine;

let prompt = TemplateEngine::new().register_template("template", "{{#if true}}Hello, world!{{/if}}");
let result = prompt.render("template").unwrap();

assert_eq!(result.to_string(), "Hello, world!".to_string());

pub fn render_context<T>(&self, name: &str, data: &T) -> Result<Box<dyn Prompt>>where T: Serialize,

Renders a Handlebars template with the given data and returns the result as a Boxed trait object.

Arguments

data - A reference to the data to be used in the template rendering.

Returns

Returns a Result containing the rendered template as a Boxed trait object.

Example

use serde_json::json;
use orca::prompt::TemplateEngine;

let prompt = TemplateEngine::new().register_template("template", "Hello, {{name}}!");
let data = json!({"name": "world"});
let result = prompt.render_context("template", &data).unwrap();

assert_eq!(result.to_string(), "Hello, world!".to_string());

pub fn render_chat<T>(&self, name: &str, data: Option<&T>) -> Result<ChatPrompt>where T: Serialize,

Renders a Handlebars template with the given data and returns the result as a vector of Messages.

Arguments

data - A reference to the data to be used in the template rendering.

Returns

Returns a Result containing the rendered template as a vector of Messages if successful, or an error if the rendering fails.

Example

use serde_json::json;
use orca::prompt::TemplateEngine;
use orca::prompt::chat::{Role, Message};

let prompt = TemplateEngine::new().register_template("template", "{{#system}}Hello, {{name}}!{{/system}}");
let data = json!({"name": "world"});
let result = prompt.render_chat("template", Some(&data));
assert_eq!(result.unwrap(), vec![Message::new(Role::System, "Hello, world!")]);