Trait orca::prompt::Prompt

pub trait Prompt: Sync + Send {
    // Required methods
    fn save(&mut self, data: Box<dyn Prompt>) -> Result<()>;
    fn to_string(&self) -> Result<String>;
    fn as_str(&self) -> Result<&str>;
    fn to_chat(&self) -> Result<ChatPrompt>;
    fn clone_prompt(&self) -> Box<dyn Prompt>;
}

A trait representing a prompt for a Large Language Model.

The Prompt trait provides methods to transform, clone, save, and represent prompts in various formats.

Required Methods

fn save(&mut self, data: Box<dyn Prompt>) -> Result<()>

Save the data from another Prompt into the current one.

Arguments

data - A boxed trait object implementing the Prompt trait.

Returns

Result<()> - An empty result indicating success or an error.

Examples

use orca::prompt;
use orca::prompt::Prompt;

let mut my_prompt = prompt!("Some prompt");
let another_prompt = prompt!("Some other prompt");
my_prompt.save(another_prompt).unwrap();

fn to_string(&self) -> Result<String>

Convert the current prompt to a String.

Returns

Result<String> - The String representation of the prompt or an error.

Examples

use orca::prompt;
use orca::prompt::Prompt;

let my_prompt = prompt!("Some prompt");
assert_eq!(my_prompt.to_string(), "Some prompt".to_string());

fn as_str(&self) -> Result<&str>

Get the current prompt as a string slice.

Returns

Result<&str> - The string slice representation of the prompt or an error.

fn to_chat(&self) -> Result<ChatPrompt>

Convert the current prompt to a ChatPrompt.

Returns

Result<ChatPrompt> - The ChatPrompt representation of the prompt or an error.

fn clone_prompt(&self) -> Box<dyn Prompt>

Clone the current prompt into a Boxed trait object.

Returns

Box<dyn Prompt> - The cloned prompt.

Examples

use orca::prompt;
use orca::prompt::Prompt;

let my_prompt = prompt!("Some prompt");
let cloned_prompt = my_prompt.clone_prompt();