Struct orca::record::Record

pub struct Record {
    pub header: Option<String>,
    pub content: Content,
    pub metadata: Option<String>,
}

Fields
header: Option<String>

Header information for the record.

content: Content

Content of the record.

metadata: Option<String>

Metadata for the record (present in PDFs, for example).

Implementations

impl Record

pub fn new(content: Content) -> Record

Create a new record with the given content.

pub fn with_header(self, header: String) -> Self

Modify the header of the record.

pub fn with_metadata(self, metadata: String) -> Self

Modify the metadata of the record.

pub fn with_content(self, content: Content) -> Self

pub fn split(&self, max_tokens: usize) -> Vec<Record>

Splits the content of a Record into multiple smaller records based on character count.

This function divides the content of a Record into smaller chunks of approximately equal size. The chunks are determined by the maximum number of characters allowed per chunk. If the content is a vector of strings, each string will be split into chunks separately.

Arguments

chunks - The desired number of chunks the content should be split into.

Returns

A vector of Record where each record contains a chunk of the original content.

Example

let record = Record::new(Content::String("Hello World".into()));
let records = record.split(5);
assert_eq!(records.len(), 2);

pub fn split_with_tokenizer( &self, chunks: usize, tokenizer: Tokenizer ) -> Result<Vec<Record>>

Splits the content of a Record into multiple smaller records using a tokenizer.

This function divides the content of a Record into smaller chunks using a specified tokenizer. The chunks are determined by the tokenizer and the number of chunks parameter. If the content is a vector of strings, each string will be split into chunks separately.

Arguments

chunks - The desired number of chunks the content should be split into.

tokenizer - The tokenizer to be used for splitting the content. This can be a Huggingface tokenizer, a tokenizer from a file, or a tokenizer from bytes.

Returns

A Result containing a vector of Record where each record contains a chunk of the original content, or an error if there was a problem with tokenization.

Example

let record = Record::new(Content::String("Hello World".into()));
let records = record.split_with_tokenizer(2, Tokenizer::Huggingface("path_to_tokenizer".into())).unwrap();
assert_eq!(records.len(), 2);